refactor query related to `src/typings/query.ts`, change startQuery into object parameter signatures
